pyproject.toml
[project]
name = "budget-app"
version = "0.1.0"
description = "Users can create and edit budgets for different durations (monthly/annually)."
authors = [
    {name = "Audrey Theriault-Allaire",email = "audrey2699@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]

[tool.poetry]
packages = [{include = "budget_app", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
